chore: rename `model` folder to `models`
